refactor: move custom style update logic to a dedicated hook
